RFM stands for Recency, frequency and monetary analysis
RFM stands for Recency, frequency and monetary analysis. It is used in segmentation. The Frequency value is derived by checking the frequency of skills in the project, Monetary value from the bid value and recency value from the days left in the dataset. An interactive dash board is created which allows viewing the skills in each segment, quartile and their frequency value